requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alpha-vantage==2.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
appdirs==1.4.4
appnope==0.1.4
arxiv==2.1.3
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.1
Brotli==1.1.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.1
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.12.1
dataclasses-json==0.6.7
db-dtypes==1.2.0
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.1
emoji==2.14.0
eval_type_backport==0.2.0
exchange_calendars==4.5.4
executing==2.0.1
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
feedparser==6.0.11
filelock==3.15.4
filetype==1.2.0
fonttools==4.51.0
fredapi==0.5.2
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gpt-researcher==0.9.6
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
htmldocx==0.0.6
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.6.0
joblib==1.4.0
json5==0.9.25
json_repair==0.29.10
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.132
loguru==0.7.2
lxml==5.2.2
lxml_html_clean==0.2.2
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.8.4
matplotlib-inline==0.1.7
md2pdf==1.0.1
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.51.1
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pandas_market_calendars==4.4.0
parse==1.20.1
parso==0.8.4
patsy==0.5.6
peewee==3.17.5
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.0
pydyf==0.11.0
pyee==11.1.0
Pygments==2.17.2
pyluach==2.2.0
PyMuPDF==1.24.11
pyparsing==3.1.2
pypdf==5.0.1
pyphen==0.16.0
pyppeteer==2.0.0
pyquery==2.0.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
RapidFuzz==3.10.0
referencing==0.35.0
regex==2024.9.11
requests==2.32.3
requests-html==0.10.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
streamlit==1.33.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
torch==2.4.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
unstructured==0.15.13
unstructured-client==0.26.0
urllib3==1.26.18
uvicorn==0.30.1
uvloop==0.19.0
w3lib==2.1.2
watchfiles==0.22.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
websockets==10.4
wrapt==1.16.0
yahoo-fin==0.8.9.1
yarl==1.9.4
yc-central==0.0.1
yfinance==0.2.40
zipp==3.19.1
zopfli==0.2.3
